Add manual, and automatic rescheduling
