+++ lib/butler_web/live/todo_live/index.html.leex
@@ -49,7 +49,8 @@
                 </div>
 
                 <button type="button"
-                    class="inline-flex items-center p-3 border border-transparent shadow-sm text-sm leading-4 font-medium rounded-full text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
+                    phx-click="run_scheduler"
+                    class="inline-flex items-center p-3 border border-indigo-500 shadow-sm text-sm leading-4 font-medium rounded-full text-indigo-600 bg-white hover:bg-indigo-700 hover:text-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
                     <svg class="h-4 w-4" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
                         stroke="currentColor">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
