+++ lib/butler_web/router.ex
@@ -30,8 +30,8 @@ defmodule ButlerWeb.Router do
     live "/todos/new", TodoLive.Index, :new
     live "/todos/:id/edit", TodoLive.Index, :edit
 
-    live "/todos/:id", TodoLive.Show, :show
-    live "/todos/:id/show/edit", TodoLive.Show, :edit
+    # live "/todos/:id", TodoLive.Show, :show
+    # live "/todos/:id/show/edit", TodoLive.Show, :edit
   end
 
   # Other scopes may use custom stacks.
