+++ lib/butler_web/templates/layout/live.html.leex
@@ -1,11 +1,11 @@
-<main role="main" class="">
-  <p class="alert alert-info" role="alert"
-    phx-click="lv:clear-flash"
-    phx-value-key="info"><%= live_flash(@flash, :info) %></p>
+<main role="main">
+  <%= if live_flash(@flash, :info) do %>
+    <%= render "_success_notif.html", message: live_flash(@flash, :info) %>
+  <% end %>
 
-  <p class="alert alert-danger" role="alert"
-    phx-click="lv:clear-flash"
-    phx-value-key="error"><%= live_flash(@flash, :error) %></p>
+  <%= if live_flash(@flash, :error) do %>
+    <%= render "_error_notif.html", message: live_flash(@flash, :error) %>
+  <% end %>
 
   <%= @inner_content %>
 </main>
